Add k8s deployment manifest and podman info
This commit adds a k8s manifest in deployment.yaml.  This manifest can
be used to deploy the container (pulling from github) to k8s or you can
use podman.  The Makefile was updated with targets for podman to deploy
and destroy.  It's nice to use podman when you don't want a full blown
k8s deployment.